базовая проверка ячейки вынесена отдельно
