+++ src/main/java/svkreml/fun/MyWebDriver.java
@@ -1,21 +1,17 @@
-package svkreml.fun.sudoku;
+package svkreml.fun;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.remote.CapabilityType;
-import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.HashMap;
 
 public class MyWebDriver {
     WebDriverWait wait;
     WebDriver chromeDriver;
-    MyWebDriver(){
+
+    MyWebDriver() {
 
         System.setProperty("webdriver.chrome.driver", "chromedriver.exe");
         HashMap<String, Object> chromePrefs = new HashMap<String, Object>();
@@ -26,9 +22,8 @@ public class MyWebDriver {
         options.setExperimentalOption("prefs", chromePrefs);
 
 
-
         chromeDriver = new ChromeDriver(options);
-        wait= new WebDriverWait(chromeDriver, 30);
+        wait = new WebDriverWait(chromeDriver, 30);
 
     }
