+++ src/main/java/svkreml/fun/Solver.java
@@ -1,16 +1,18 @@
-package svkreml.fun.sudoku;
+package svkreml.fun;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
+import svkreml.fun.sudoku.Field;
+import svkreml.fun.sudoku.SodokuSolver;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
 
 public class Solver {
-    public static void main(String[] args) throws InterruptedException {
+    public static void main(String[] args) {
         MyWebDriver myWebDriver = null;
         try {
             myWebDriver = new MyWebDriver();
@@ -18,7 +20,7 @@ public class Solver {
 
             myWebDriver.get("https://www.websudoku.com/?level=2");
             Thread.sleep(1000);
-            WebDriver frame = myWebDriver.chromeDriver.switchTo().frame(0);
+            myWebDriver.chromeDriver.switchTo().frame(0);// ервый взгляд ничего не проиходит, но содержимое странице во фрейме и без этой строчки содержимое фрейма недоступно
             for (int i = 0; i < 10; i++) {
                 loop(myWebDriver);
             }
@@ -35,7 +37,6 @@ public class Solver {
         List<WebElement> elements = myWebDriver.chromeDriver.findElements(By.xpath("//table[@id='puzzle_grid']/tbody/tr/td/input"));
 
         Field field = new Field();
-
 
 
         Iterator<WebElement> iterator = elements.iterator();
@@ -67,10 +68,10 @@ public class Solver {
             }
         }
         try {
-        Thread.sleep(2000);
-        myWebDriver.wait.until(ExpectedConditions.presenceOfElementLocated(By.xpath("//input[@value=' How am I doing? ']")));
-        myWebDriver.chromeDriver.findElement(By.xpath("//input[@value=' How am I doing? ']")).click();
-        Thread.sleep(2000);
+            Thread.sleep(2000);
+            myWebDriver.wait.until(ExpectedConditions.presenceOfElementLocated(By.xpath("//input[@value=' How am I doing? ']")));
+            myWebDriver.chromeDriver.findElement(By.xpath("//input[@value=' How am I doing? ']")).click();
+            Thread.sleep(2000);
 
             myWebDriver.wait.until(ExpectedConditions.presenceOfElementLocated(By.xpath("//input[@value=' Bring on a new puzzle! ']")));
             myWebDriver.chromeDriver.findElement(By.xpath("//input[@value=' Bring on a new puzzle! ']")).click();
