+++ src/main/java/svkreml/fun/sudoku/Field.java
@@ -1,21 +1,28 @@
 package svkreml.fun.sudoku;
 
+import java.util.InputMismatchException;
+
 public class Field {
-    Integer[][] field = new Integer[9 + 1][9 + 1];
-    public  Field(String input) {
+    private Integer[][] field = new Integer[9 + 1][9 + 1];
+
+    public Field(String input) {
         input = input.replaceAll("[^1234567890\\-]{1}", "");
+        if (input.length() != 81) throw new InputMismatchException("должна быть ровно 81 цифра");
         for (int y = 1; y <= 9; y++) {
             for (int x = 1; x <= 9; x++) {
                 char c = input.charAt((y - 1) * 9 + x - 1);
                 if (c == '-') continue;
-                //if (line.charAt(x) >= '1' && line.charAt(x) <= '9') throw new InputMismatchException();
                 field[x][y] = c - '0';
             }
         }
     }
 
     public Field() {
+    }
 
+    public static int getOffset(Integer i) {
+        int j = i - 1;
+        return j - j % 3;
     }
 
     public void setCell(int x, int y, int value) {
@@ -40,16 +47,16 @@ public class Field {
                 if (x % 3 == 0) System.out.print("|");
                 else System.out.print(" ");
             }
-
             System.out.println();
             if (y % 3 == 0) System.out.println("¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯");
         }
         return solvedField;
     }
 
-    public Integer getCell(int x, int y) {
+    public Integer getCell(Integer x, Integer y) {
         if (!((y >= 1) && (y <= 9))) throw new IndexOutOfBoundsException();
         if (!((x >= 1) && (x <= 9))) throw new IndexOutOfBoundsException();
+
         return field[x][y];
     }
 }
